clear pydantic, some tests pass
