+++ src/caustics/sims/__init__.py
@@ -1,4 +1,5 @@
 from .lens_source import LensSource
 from .microlens import Microlens
+from .simulator import build_simulator
 
-__all__ = ("LensSource", "Microlens")
+__all__ = ("LensSource", "Microlens", "build_simulator")
