+++ tests/test_simulator_runs.py
@@ -9,81 +9,79 @@ from caustics.light import Sersic
 from caustics.utils import gaussian
 from caustics import build_simulator
 
-from utils import mock_from_file
-
 
 def test_simulator_runs(sim_source, device, mocker):
     if sim_source == "yaml":
-        yaml_str = """\
-        cosmology: &cosmology
-            name: "cosmo"
-            kind: FlatLambdaCDM
+        # yaml_str = """\
+        # cosmology: &cosmology
+        #     name: "cosmo"
+        #     kind: FlatLambdaCDM
 
-        lensmass: &lensmass
-            name: lens
-            kind: SIE
-            params:
-                z_l: 1.0
-                x0: 0.0
-                y0: 0.01
-                q: 0.5
-                phi: pi / 3.0
-                b: 1.0
-            init_kwargs:
-                cosmology: *cosmology
+        # lensmass: &lensmass
+        #     name: lens
+        #     kind: SIE
+        #     params:
+        #         z_l: 1.0
+        #         x0: 0.0
+        #         y0: 0.01
+        #         q: 0.5
+        #         phi: pi / 3.0
+        #         b: 1.0
+        #     init_kwargs:
+        #         cosmology: *cosmology
 
-        source: &source
-            name: source
-            kind: Sersic
-            params:
-                x0: 0.01
-                y0: -0.03
-                q: 0.6
-                phi: -pi / 4
-                n: 1.5
-                Re: 0.5
-                Ie: 1.0
+        # source: &source
+        #     name: source
+        #     kind: Sersic
+        #     params:
+        #         x0: 0.01
+        #         y0: -0.03
+        #         q: 0.6
+        #         phi: -pi / 4
+        #         n: 1.5
+        #         Re: 0.5
+        #         Ie: 1.0
 
-        lenslight: &lenslight
-            name: lenslight
-            kind: Sersic
-            params:
-                x0: 0.0
-                y0: 0.01
-                q: 0.7
-                phi: pi / 4
-                n: 3.0
-                Re: 0.7
-                Ie: 1.0
+        # lenslight: &lenslight
+        #     name: lenslight
+        #     kind: Sersic
+        #     params:
+        #         x0: 0.0
+        #         y0: 0.01
+        #         q: 0.7
+        #         phi: pi / 4
+        #         n: 3.0
+        #         Re: 0.7
+        #         Ie: 1.0
 
-        psf: &psf
-            func: caustics.utils.gaussian
-            kwargs:
-                pixelscale: 0.05
-                nx: 11
-                ny: 11
-                sigma: 0.2
-                upsample: 2
+        # psf: &psf
+        #     func: caustics.utils.gaussian
+        #     kwargs:
+        #         pixelscale: 0.05
+        #         nx: 11
+        #         ny: 11
+        #         sigma: 0.2
+        #         upsample: 2
 
-        simulator:
-            name: simulator
-            kind: LensSource
-            params:
-                z_s: 2.0
-            init_kwargs:
-                # Single lens
-                lens: *lensmass
-                source: *source
-                lens_light: *lenslight
-                pixelscale: 0.05
-                pixels_x: 50
-                psf: *psf{quad_level}
-        """
-        mock_from_file(
-            mocker, yaml_str.format(quad_level="")
-        )  # fixme, yaml should be able to accept None
+        # simulator:
+        #     name: simulator
+        #     kind: LensSource
+        #     params:
+        #         z_s: 2.0
+        #     init_kwargs:
+        #         # Single lens
+        #         lens: *lensmass
+        #         source: *source
+        #         lens_light: *lenslight
+        #         pixelscale: 0.05
+        #         pixels_x: 50
+        #         psf: *psf{quad_level}
+        # """
+        # mock_from_file(
+        #     mocker, yaml_str.format(quad_level="")
+        # )  # fixme, yaml should be able to accept None
         sim = build_simulator("/path/to/sim.yaml")  # Path doesn't actually exists
-        mock_from_file(mocker, yaml_str.format(quad_level="\n        quad_level: 3"))
+        # mock_from_file(mocker, yaml_str.format(quad_level="\n        quad_level: 3"))
         sim_q3 = build_simulator("/path/to/sim.yaml")  # Path doesn't actually exists
     else:
         # Model
@@ -147,7 +145,6 @@ def test_simulator_runs(sim_source, device, mocker):
     assert torch.all(
         torch.isfinite(
             sim(
-                {},
                 source_light=True,
                 lens_light=True,
                 lens_source=True,
@@ -158,7 +155,6 @@ def test_simulator_runs(sim_source, device, mocker):
     assert torch.all(
         torch.isfinite(
             sim(
-                {},
                 source_light=True,
                 lens_light=True,
                 lens_source=False,
@@ -169,7 +165,6 @@ def test_simulator_runs(sim_source, device, mocker):
     assert torch.all(
         torch.isfinite(
             sim(
-                {},
                 source_light=True,
                 lens_light=False,
                 lens_source=True,
@@ -180,7 +175,6 @@ def test_simulator_runs(sim_source, device, mocker):
     assert torch.all(
         torch.isfinite(
             sim(
-                {},
                 source_light=False,
                 lens_light=True,
                 lens_source=True,
